rename Typesafe -> Lightbend in actual code
