+++ build.sbt
@@ -36,11 +36,11 @@ lazy val root = Project(
     aggregate = Seq(core, amazon, github, cli, jenkins, server)
 ) dependsOn(gui) // this does the actual aggregation
 
-lazy val core    = project settings (deps: _*)
-lazy val github  = project dependsOn (core)
-lazy val jenkins = project dependsOn (core)
-lazy val typesafe = project dependsOn (core)
-lazy val cli     = project dependsOn (github)
-lazy val amazon  = project dependsOn (core) settings (amazonDeps: _*)
-lazy val server  = project dependsOn (amazon, github, jenkins, typesafe)
-lazy val gui     = project dependsOn (server) enablePlugins(PlayScala) settings (guiSettings: _*)
+lazy val core      = project settings (deps: _*)
+lazy val github    = project dependsOn (core)
+lazy val jenkins   = project dependsOn (core)
+lazy val lightbend = project dependsOn (core)
+lazy val cli       = project dependsOn (github)
+lazy val amazon    = project dependsOn (core) settings (amazonDeps: _*)
+lazy val server    = project dependsOn (amazon, github, jenkins, lightbend)
+lazy val gui       = project dependsOn (server) enablePlugins(PlayScala) settings (guiSettings: _*)
