+++ gui/app/controllers/Dashboard.scala
@@ -7,7 +7,7 @@ import scabot.github._
 import scabot.jenkins.JenkinsService
 import scabot.core
 import scabot.server.Actors
-import scabot.typesafe.TypesafeApi
+import scabot.lightbend.LightbendApi
 import spray.json._
 import scala.util._
 import scala.concurrent.Future
@@ -20,7 +20,7 @@ object Dashboard {
       reviewed: String)
 }
 
-class Dashboard @Inject() (val system: ActorSystem) extends Controller with GithubService with JenkinsService with TypesafeApi with core.Configuration with core.HttpClient with Actors {
+class Dashboard @Inject() (val system: ActorSystem) extends Controller with GithubService with JenkinsService with LightbendApi with core.Configuration with core.HttpClient with Actors {
   import Dashboard._
 
   override def configFile = new java.io.File(sys.props("scabot.config.file"))
