+++ lightbend/src/main/scala/scabot/lightbend/LightbendApi.scala
@@ -1,21 +1,21 @@
 package scabot
-package typesafe
+package lightbend
 
 import spray.json.{RootJsonFormat, DefaultJsonProtocol}
 
-trait TypesafeApi extends TypesafeApiTypes with DefaultJsonProtocol with TypesafeApiActions
+trait LightbendApi extends LightbendApiTypes with DefaultJsonProtocol with LightbendApiActions
 
-trait TypesafeApiTypes {
+trait LightbendApiTypes {
   case class CLARecord(user: String, signed: Boolean, version: Option[String], currentVersion: String)
 }
 
-trait TypesafeJsonProtocol extends TypesafeApiTypes with DefaultJsonProtocol {
+trait LightbendJsonProtocol extends LightbendApiTypes with DefaultJsonProtocol {
   private type RJF[x] = RootJsonFormat[x]
   implicit lazy val _fmtCLARecord: RJF[CLARecord] = jsonFormat4(CLARecord)
 }
 
-trait TypesafeApiActions extends TypesafeJsonProtocol with core.HttpClient {
-  class TypesafeConnection {
+trait LightbendApiActions extends LightbendJsonProtocol with core.HttpClient {
+  class LightbendConnection {
     import spray.http.{GenericHttpCredentials, Uri}
     import spray.httpx.SprayJsonSupport._
     import spray.client.pipelining._
